+++ localeyes/src/app/components/profile/profile.component.ts
@@ -1,7 +1,7 @@
 import { Component, inject } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
 
 import { PostService } from '../../services/post.service';
 import { CustomResponse, User } from '../../modals/modals';
@@ -17,6 +17,7 @@ export class ProfileComponent {
   postService = inject(PostService);
   userService = inject(UserService);
   messageService = inject(MessageService);
+  confirmationService = inject(ConfirmationService);
   router = inject(Router);
 
   ngOnInit() {
@@ -51,21 +52,34 @@ export class ProfileComponent {
   }
 
   deactivate() {
-    this.userService.deactivate().subscribe({
-      next: () => {
-        this.messageService.add({
-          severity: 'success',
-          detail: 'Deactivated user'
+    this.confirmationService.confirm({
+      message: 'Are you sure that you want to proceed?',
+      header: 'Confirmation',
+      icon: 'pi pi-exclamation-triangle',
+      acceptIcon:"None",
+      rejectIcon:"None",
+      rejectButtonStyleClass:"p-button-text",
+      accept: () => {
+        this.userService.deactivate().subscribe({
+          next: () => {
+            this.messageService.add({
+              severity: 'success',
+              detail: 'Deactivated user'
+            })
+            localStorage.clear();
+            this.router.navigate(['/login'])
+          },
+          error: () => {
+            this.messageService.add({
+              severity: 'error',
+              detail: 'Some issue at our end'
+            })
+          }
         })
-        localStorage.clear();
-        this.router.navigate(['/login'])
       },
-      error: () => {
-        this.messageService.add({
-          severity: 'error',
-          detail: 'Some issue at our end'
-        })
+      reject: () => {
+        this.messageService.add({ severity: 'error', summary: 'Rejected', detail: 'You have rejected to deactivate', life: 3000 });
       }
-    })
+    });
   }
 }
