+++ localeyes/src/app/services/question.service.ts
@@ -3,7 +3,7 @@ import { CustomResponse, NewAnswer, NewQuestion, Question } from "../modals/moda
 import { HttpClient } from "@angular/common/http";
 import { PostService } from "./post.service";
 import { AddAnswer, AddQuestion, GetPostQuestions } from "../constants/urls";
-import { Observable } from "rxjs";
+import { Observable, throwError } from "rxjs";
 
 @Injectable({
   providedIn: 'root'
@@ -20,18 +20,36 @@ export class QuestionService {
   postService = inject(PostService);
 
   addQuestion(question: NewQuestion): Observable<CustomResponse>{
+    if(!this.postService.activePost()){
+      return throwError(() => new Error('No Post Found'))
+    }
     return this.httpClient.post<CustomResponse>(AddQuestion(this.postService.activePost()?.post_id as string),question)
   }
 
   addAnswer(answer: NewAnswer): Observable<CustomResponse> {
+    if(!this.postService.activePost()){
+      return throwError(() => new Error('No Post Found'))
+    }
+    if(!this.activeQuestion()){
+      return throwError(() => new Error('No Question Found'))
+    }
     return this.httpClient.put<CustomResponse>(AddAnswer(this.postService.activePost()?.post_id as string,this.activeQuestion()?.question_id as string),answer)
   }
 
   deleteQuestion(): Observable<CustomResponse> {
+    if(!this.postService.activePost()){
+      return throwError(() => new Error('No Post Found'))
+    }
+    if(!this.activeQuestion()){
+      return throwError(() => new Error('No Question Found'))
+    }
     return this.httpClient.delete<CustomResponse>(AddAnswer(this.postService.activePost()?.post_id as string,this.activeQuestion()?.question_id as string))
   }
 
   getAllquestions(): Observable<CustomResponse> {
+    if(!this.postService.activePost()){
+      return throwError(() => new Error('No Post Found'))
+    }
     return this.httpClient.get<CustomResponse>(GetPostQuestions(this.postService.activePost()?.post_id as string))
   }
 }
